# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthStudentProfileResponse": {
        "properties": {
          "academic_status": {
            "title": "Academic Status",
            "type": "string"
          },
          "batch": {
            "title": "Batch",
            "type": "string"
          },
          "current_semester": {
            "title": "Current Semester",
            "type": "integer"
          },
          "department": {
            "title": "Department",
            "type": "string"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          },
          "onboarding_completed": {
            "title": "Onboarding Completed",
            "type": "boolean"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "program": {
            "title": "Program",
            "type": "string"
          },
          "registration_number": {
            "title": "Registration Number",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          }
        },
        "required": [
          "registration_number",
          "department",
          "program",
          "batch",
          "current_semester",
          "section",
          "academic_status",
          "is_verified",
          "preferred_name",
          "onboarding_completed"
        ],
        "title": "AuthStudentProfileResponse",
        "type": "object"
      },
...
      "FacultyAssignmentCreate": {
        "additionalProperties": false,
        "properties": {
          "course_code": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Course Code",
            "type": "string"
          },
          "faculty_user_id": {
            "exclusiveMinimum": 0.0,
            "title": "Faculty User Id",
            "type": "integer"
          },
          "section": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Section",
            "type": "string"
          },
          "semester": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Semester",
            "type": "string"
          },
          "term_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term Id"
          }
        },
        "required": [
          "faculty_user_id",
          "course_code",
          "section",
          "semester"
        ],
        "title": "FacultyAssignmentCreate",
        "type": "object"
      },
      "FacultyAssignmentResponse": {
        "properties": {
          "course_code": {
            "title": "Course Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By User Id"
          },
          "faculty_email": {
            "title": "Faculty Email",
            "type": "string"
          },
          "faculty_name": {
            "title": "Faculty Name",
            "type": "string"
          },
          "faculty_user_id": {
            "title": "Faculty User Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "semester": {
            "title": "Semester",
            "type": "string"
          },
          "term_id": {
            "title": "Term Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "term_id",
          "faculty_user_id",
          "faculty_name",
          "faculty_email",
          "course_code",
          "section",
          "semester",
          "created_by_user_id",
          "created_at"
        ],
        "title": "FacultyAssignmentResponse",
        "type": "object"
      },
      "FacultyDirectoryEntryResponse": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "full_name",
          "email"
        ],
        "title": "FacultyDirectoryEntryResponse",
        "type": "object"
      },
      "FacultyDirectoryResponse": {
        "properties": {
          "faculty": {
            "items": {
              "$ref": "#/components/schemas/FacultyDirectoryEntryResponse"
            },
            "title": "Faculty",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "faculty",
          "total",
          "offset",
          "limit"
        ],
        "title": "FacultyDirectoryResponse",
        "type": "object"
      },
      "FacultyProvisionCreate": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 5,
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Full Name",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "temporary_password": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporary Password"
          }
        },
        "required": [
          "full_name",
          "email"
        ],
        "title": "FacultyProvisionCreate",
        "type": "object"
      },
      "FacultyProvisionResponse": {
        "properties": {
          "faculty": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "temporary_password": {
            "title": "Temporary Password",
            "type": "string"
          }
        },
        "required": [
          "faculty",
          "temporary_password"
        ],
        "title": "FacultyProvisionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TimetableEntryCreate": {
        "additionalProperties": false,
        "properties": {
          "class_type": {
            "default": "lecture",
            "enum": [
              "lecture",
              "lab",
              "tutorial",
              "online",
              "hybrid",
              "other"
            ],
            "title": "Class Type",
            "type": "string"
          },
          "course_code": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Code"
          },
          "course_name": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "day": {
            "enum": [
              "Monday",
              "Tuesday",
              "Wednesday",
              "Thursday",
              "Friday",
              "Saturday",
              "Sunday"
            ],
            "title": "Day",
            "type": "string"
          },
          "end_time": {
            "title": "End Time",
            "type": "string"
          },
          "entry_kind": {
            "default": "course",
            "enum": [
              "course",
              "special_event"
            ],
            "title": "Entry Kind",
            "type": "string"
          },
          "faculty": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty"
          },
          "raw_text": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "room": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room"
          },
          "section": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "semester": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "source": {
            "default": "manual",
            "enum": [
              "manual",
              "csv",
              "xlsx",
              "docx",
              "pdf",
              "image"
            ],
            "title": "Source",
            "type": "string"
          },
          "start_time": {
            "title": "Start Time",
            "type": "string"
          }
        },
        "required": [
          "day",
          "start_time",
          "end_time"
        ],
        "title": "TimetableEntryCreate",
        "type": "object"
      },
      "TimetableEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "class_type": {
            "default": "lecture",
            "enum": [
              "lecture",
              "lab",
              "tutorial",
              "online",
              "hybrid",
              "other"
            ],
            "title": "Class Type",
            "type": "string"
          },
          "course_code": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Code"
          },
          "course_name": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "day": {
            "enum": [
              "Monday",
              "Tuesday",
              "Wednesday",
              "Thursday",
              "Friday",
              "Saturday",
              "Sunday"
            ],
            "title": "Day",
            "type": "string"
          },
          "end_time": {
            "title": "End Time",
            "type": "string"
          },
          "entry_kind": {
            "default": "course",
            "enum": [
              "course",
              "special_event"
            ],
            "title": "Entry Kind",
            "type": "string"
          },
          "faculty": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "raw_text": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "room": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room"
          },
          "section": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "semester": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "source": {
            "default": "manual",
            "enum": [
              "manual",
              "csv",
              "xlsx",
              "docx",
              "pdf",
              "image"
            ],
            "title": "Source",
            "type": "string"
          },
          "start_time": {
            "title": "Start Time",
            "type": "string"
          },
          "term_id": {
            "title": "Term Id",
            "type": "integer"
          }
        },
        "required": [
          "day",
          "start_time",
          "end_time",
          "id",
          "term_id"
        ],
        "title": "TimetableEntryResponse",
        "type": "object"
      },
...
      "UserResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "student_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthStudentProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active",
          "must_change_password",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "student",
          "faculty",
          "coordinator",
          "admin"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "UniTime AI API",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/faculty-assignments": {
      "get": {
        "operationId": "get_managed_faculty_assignments_faculty_assignments_get",
        "parameters": [
          {
            "in": "query",
            "name": "faculty_user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Faculty User Id"
            }
          },
          {
            "in": "query",
            "name": "term_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Term Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FacultyAssignmentResponse"
                  },
                  "title": "Response Get Managed Faculty Assignments Faculty Assignments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Managed Faculty Assignments",
        "tags": [
          "Faculty Assignments"
        ]
      },
      "post": {
        "operationId": "add_faculty_assignment_faculty_assignments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyAssignmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyAssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Faculty Assignment",
        "tags": [
          "Faculty Assignments"
        ]
      }
    },
    "/faculty-assignments/{assignment_id}": {
      "delete": {
        "operationId": "remove_faculty_assignment_faculty_assignments__assignment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Faculty Assignment",
        "tags": [
          "Faculty Assignments"
        ]
      }
    },
    "/faculty-directory": {
      "get": {
        "operationId": "get_faculty_directory_faculty_directory_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDirectoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Faculty Directory",
        "tags": [
          "Faculty Directory"
        ]
      },
      "post": {
        "operationId": "create_faculty_account_faculty_directory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyProvisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyProvisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Faculty Account",
        "tags": [
          "Faculty Directory"
        ]
      }
    },
    "/faculty/assignments": {
      "get": {
        "operationId": "get_my_faculty_assignments_faculty_assignments_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FacultyAssignmentResponse"
                  },
                  "title": "Response Get My Faculty Assignments Faculty Assignments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Faculty Assignments",
        "tags": [
          "Faculty"
        ]
      }
    },
    "/faculty/timetable": {
      "get": {
        "operationId": "get_my_faculty_timetable_faculty_timetable_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimetableEntryResponse"
                  },
                  "title": "Response Get My Faculty Timetable Faculty Timetable Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Faculty Timetable",
        "tags": [
          "Faculty"
        ]
      }
    },
...
    "/timetable": {
      "get": {
        "operationId": "get_timetable_timetable_get",
        "parameters": [
          {
            "in": "query",
            "name": "term_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Term Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimetableEntryResponse"
                  },
                  "title": "Response Get Timetable Timetable Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Timetable"
      },
      "post": {
        "operationId": "create_timetable_entry_timetable_post",
        "parameters": [
          {
            "in": "query",
            "name": "term_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Term Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Timetable Entry"
      }
    },
...
    "/timetable/{entry_id}": {
      "delete": {
        "operationId": "delete_timetable_entry_timetable__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Timetable Entry"
      },
      "get": {
        "operationId": "get_timetable_entry_timetable__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Timetable Entry"
      }
    },
...
  }
}